changes unsocketed
replaced calling blob directly for post, get delete requests, created endpoints to handle this information, recreated post management, which needs some additional testing. as of now it can post one or more items, click on the image which pops up the model, once the model is opened the user can click out. post management has the delete button within the model, a fetch request is made with the title to get the itemID from the Database, that item ID is used to delete the rows corresponding with it.  allowing for a "post" like feature removed web socket, http, left depedencies for future use.